1. update train_birds+.py 2. combine 2 baseline models as model_options

+++ model/vgg_dc.py
@@ -22,12 +22,13 @@ class VGG(nn.Module):
         if init_weights:
             self._initialize_weights()
 
-    def drop_channel_block(self, x, drop_rate=0.1):
+    ### strategy1: cal correlation based on Euclidean distance ###
+    def drop_channel_block_s1(self, x, drop_rate=0.2):
         """
         x: bs*c*h*w
         """
         if self.training:
-            ### generate adj matrix
+            ### generate correlation matrix
             bs, c, h, w = x.size()
             mask = nn.AvgPool2d(kernel_size=3, stride=1, padding=1)(x) #bs*c*h*w
             mask = mask.view(bs,c,-1)
@@ -40,22 +41,69 @@ class VGG(nn.Module):
             dot_product = torch.bmm(loc_xy, torch.transpose(loc_xy, 1, 2))
 
             distances = mask_square.unsqueeze(2) - 2.0*dot_product + mask_square.unsqueeze(1)
-            distances = -distances
+            distances = torch.sqrt(distances)
 
+            distances = -distances
             d_min = torch.min(distances, dim=-1, keepdim=True)[0]
             d_max = torch.max(distances, dim=-1, keepdim=True)[0]
-
             adjs = (distances - d_min) / (d_max - d_min)
-
             ### drop function
             topk = int(c*drop_rate)
             value, indices = torch.topk(adjs, topk, dim = 2) # [bs, c, c] -> [bs, c, topk]
-            value = value[:,:,-1].view(bs, c,-1)
-            mask = adjs<value # [bs, c, c]
+            value = value[:,:,-1].view(bs, c, -1)
+            mask = adjs < value # [bs, c, c]
+
+            drop_index = torch.randint(low=0, high=c, size=(bs,1,1)).to(distances.device).repeat(1,1,c).long()
+            mask = torch.gather(mask, 1, drop_index).view(bs,-1,1,1).repeat(1,1,h,w).float() #[bs,c,h,w]
+
+            ### get masked heatmap and remained heatmap for visualization
+            mask_remain = mask
+            mask_drop = 1 - mask_remain
+            x_remain = x* mask_remain
+            x_drop = x* mask_drop
+            x_select_channel = torch.gather(x, 1, drop_index[:,:,0].view(bs,1,1,1).repeat(1,1,h,w))
+            ### normalize
+            x_out = x_remain / (1-drop_rate)
+
+            return x_out, torch.sum(x_remain, dim=1, keepdim=True), torch.sum(x_drop, dim=1, keepdim=True), x_select_channel
+        else:
+            return x, torch.sum(x, dim=1, keepdim=True), torch.sum(x, dim=1, keepdim=True), torch.sum(x, dim=1, keepdim=True)
+
+    ### strategy2: cal correlation based on BP ###
+    def drop_channel_block_s2(self, x, drop_rate=0.05):
+        """
+        x: bs*c*h*w
+        """
+        if self.training:
+            ### generate correlation matrix
+            bs, c, w, h = x.size()
+            x_norm = x.view(bs, c, -1)
+            x_norm = torch.nn.functional.normalize(x_norm, dim = 2)
+            adjs = torch.bmm(x_norm, torch.transpose(x_norm, 1, 2))
+
+            # mask = adjs > drop_thr
+            ### drop function
+            topk = int(c*drop_rate)
+            value, indices = torch.topk(adjs, topk, dim = 2) # [bs, c, c] -> [bs, c, topk]
+            value = value[:,:,-1].view(bs, c, -1)
+            mask = adjs < value # [bs, c, c]
+            
             drop_index = torch.randint(low=0, high=c, size=(bs,1,1)).to(adjs.device).repeat(1,1,c).long()
             mask = torch.gather(mask, 1, drop_index).view(bs,-1,1,1).repeat(1,1,h,w).float() #[bs,c,h,w]
-            # normalize & set correlated feat to 0
-            x = x* mask / (1-drop_rate)
+
+            ### get masked heatmap and remained heatmap for visualization
+            mask_remain = mask
+            mask_drop = 1 - mask_remain
+        
+            x_remain = x* mask_remain
+            x_drop = x* mask_drop
+            x_select_channel = torch.gather(x, 1, drop_index[:,:,0].view(bs,1,1,1).repeat(1,1,h,w))
+            ### normalize
+            x_out = x_remain / (1-drop_rate)
+
+            return x_out, torch.sum(x, dim=1, keepdim=True), torch.sum(x_remain, dim=1, keepdim=True), torch.sum(x_drop, dim=1, keepdim=True), x_select_channel, x[0].unsqueeze(1)
+        else:
+            return x, torch.sum(x, dim=1, keepdim=True), [], [], [], []
 
         return x
 
@@ -64,9 +112,10 @@ class VGG(nn.Module):
         x = nn.Sequential(*list(self.features.children())[:7])(x) #[64,224,224]
         # x = self.drop_channel_block(x) # add dc block v1
         x = nn.Sequential(*list(self.features.children())[7:14])(x) #[128,112,112]
-        # x = self.drop_channel_block(x) # add dc block v1
+        # x = self.drop_channel_block(x) # add dc block v2
         x = nn.Sequential(*list(self.features.children())[14:27])(x) #[256,56,56]
-        # x = self.drop_channel_block(x) # add dc block v3
+        x, heatmap_all, heatmap_remain, heatmap_drop, select_channel, all_channel = self.drop_channel_block_s1(x) # add dc block v3
+        # x, heatmap_all, heatmap_remain, heatmap_drop, select_channel, all_channel = self.drop_channel_block_s2(x) # add dc block v3
         x = nn.Sequential(*list(self.features.children())[27:40])(x) #[512,28,28]
         # x = self.drop_channel_block(x) # add dc block v4
         x = nn.Sequential(*list(self.features.children())[40:])(x) #[512,14,14]
@@ -75,7 +124,7 @@ class VGG(nn.Module):
         x = self.avgpool(x) #[512,7,7]
         x = torch.flatten(x, 1)
         x = self.cls(x)
-        return x
+        return x, heatmap_all, heatmap_remain, heatmap_drop, select_channel, all_channel
 
     def _initialize_weights(self):
         for m in self.modules():
